+++ tests/unit/user_guide_test.py
@@ -23,6 +23,7 @@ def user_guide_snippet_dynamic_scratch_var() -> pt.Expr:
 
 
 @pytest.mark.parametrize("snippet", [user_guide_snippet_dynamic_scratch_var])
+@pytest.mark.serial  # Serial due to scratch generation
 def test_user_guide_snippets(snippet):
     assert_new_v_old(snippet, 6, "user_guide")
 
@@ -110,6 +111,7 @@ USER_GUIDE_SNIPPETS_COPACETIC = [
 
 
 @pytest.mark.parametrize("snippet", USER_GUIDE_SNIPPETS_COPACETIC)
+@pytest.mark.serial  # Serial due to scratch generation
 def test_user_guide_snippets_good(snippet):
     assert_new_v_old(snippet, 6, "user_guide")
 
@@ -123,6 +125,7 @@ USER_GUIDE_SNIPPETS_ERRORING = {
 
 
 @pytest.mark.parametrize("snippet_etype_e", USER_GUIDE_SNIPPETS_ERRORING.items())
+@pytest.mark.serial  # Serial due to scratch generation
 def test_user_guide_snippets_bad(snippet_etype_e):
     snippet, etype_e = snippet_etype_e
     etype, e = etype_e